% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascvd_30y_prevent_base.R
\name{ascvd_30y_prevent_base}
\alias{ascvd_30y_prevent_base}
\title{PREVENT 30-year ASCVD risk score (base model)}
\usage{
ascvd_30y_prevent_base(
  gender = c("male", "female"),
  age,
  sbp,
  bp_med,
  totchol,
  hdl,
  statin,
  diabetes,
  smoker,
  egfr,
  bmi,
  ...
)
}
\arguments{
\item{gender}{patient gender (male, female)}

\item{age}{patient age (years), between 30 and 79}

\item{sbp}{Systolic blood pressure (mm Hg)}

\item{bp_med}{Patient is on a blood pressure medication (1=Yes, 0=No)}

\item{totchol}{Total cholesterol (mg/dL)}

\item{hdl}{HDL cholesterol (mg/dL)}

\item{statin}{Patient is on a statin (1=Yes, 0=No)}

\item{diabetes}{Diabetes (1=Yes, 0=No)}

\item{smoker}{Current smoker (1=Yes, 0=No)}

\item{egfr}{Estimated glomerular filtration rate (mL/min/1.73m2)}

\item{bmi}{Body mass index (kg/m2)}

\item{...}{Additional predictors can be passed and will be ignored}
}
\value{
30-year ASCVD risk estimate (percent)
}
\description{
Computes 30-year risk for ASCVD (atherosclerotic cardiovascular disease)
using the American Heart Association PREVENT equations (2023) base model.
}
\examples{
library(CVrisk)
ascvd_30y_prevent_base(
  gender = "female", age = 50,
  sbp = 160, bp_med = 1,
  totchol = 200, hdl = 45,
  statin = 0, diabetes = 1, smoker = 0,
  egfr = 90, bmi = 35
)

}
\references{
Khan SS, Matsushita K, Sang Y, Ballew SH, Grams ME, Surapaneni A, Blaha MJ,
Carson AP, Chang AR, Ciemins E, Go AS, Gutierrez OM, Hwang SJ, Jassal SK,
Kovesdy CP, Lloyd-Jones DM, Shlipak MG, Palaniappan LP, Sperling L,
Virani SS, Tuttle K, Neeland IJ, Chow SL, Rangaswami J, Pencina MJ,
Ndumele CE, Coresh J; Chronic Kidney Disease Prognosis Consortium and the
American Heart Association Cardiovascular-Kidney-Metabolic Science Advisory
Group. Development and Validation of the American Heart Association's
PREVENT Equations. Circulation. 2024 Feb 6;149(6):430-449.
}
