% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascvd_pooled_coef.R
\docType{data}
\name{ascvd_pooled_coef}
\alias{ascvd_pooled_coef}
\title{ascvd_pooled_coef}
\format{A data frame with 4 obs. and 17 variables:
\describe{
  \item{code}{ICD-9 or ICD-10 codes}
  \item{phecode}{Phecode}
  \item{vocabulary_id}{ICD code version, either ICD9CM or ICD10CM}
}}
\usage{
ascvd_pooled_coef
}
\description{
A dataset containing the ASCVD pooled cohort coeffecients
}
\keyword{datasets}
