% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CVrisk.R
\name{compute_CVrisk}
\alias{compute_CVrisk}
\title{Compute multiple CV risk scores}
\usage{
compute_CVrisk(
  df,
  scores = c("ascvd_10y_accaha", "ascvd_10y_frs", "ascvd_10y_frs_simple"),
  age,
  gender,
  race,
  sbp = NULL,
  bmi = NULL,
  hdl = NULL,
  totchol = NULL,
  bp_med = NULL,
  smoker = NULL,
  diabetes = NULL
)
}
\arguments{
\item{df}{input dataframe}

\item{scores}{which scores to compute}

\item{age}{patient age in years}

\item{gender}{patient gender (male or female)}

\item{race}{patient race}

\item{sbp}{character string of systolic blood pressure (in mm Hg) column}

\item{bmi}{string of bmi column}

\item{hdl}{string of hdl column}

\item{totchol}{string of totchol column}

\item{bp_med}{string of bp_med column}

\item{smoker}{string of smoker column}

\item{diabetes}{string of diabetes column}
}
\value{
input data frame with risk score results appended as columns
}
\description{
Compute multiple CV risk scores
}
\examples{

library(CVrisk)
compute_CVrisk(sample_data, age = "age", race = "race", gender = "gender", bmi = "BMI", sbp = "sbp", 
            hdl = "hdl", totchol = "totchol", bp_med = "bp_med", smoker = "smoker", diabetes="diabetes")

}
