% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVrisk-package.R
\docType{package}
\name{CVrisk-package}
\alias{CVrisk}
\alias{CVrisk-package}
\title{CVrisk: Compute Risk Scores for Cardiovascular Diseases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate various cardiovascular disease risk scores from the Framingham Heart Study (FHS), the American College of Cardiology (ACC), and the American Heart Association (AHA) as described in D’agostino, et al (2008) \doi{10.1161/circulationaha.107.699579}, Goff, et al (2013) \doi{10.1161/01.cir.0000437741.48606.98}, and Mclelland, et al (2015) \doi{10.1016/j.jacc.2015.08.035}, and Khan, et al (2024) \doi{10.1161/CIRCULATIONAHA.123.067626}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vcastro.github.io/CVrisk/}
  \item \url{https://github.com/vcastro/CVrisk/}
  \item Report bugs at \url{https://github.com/vcastro/CVrisk/issues}
}

}
\author{
\strong{Maintainer}: Victor Castro \email{vcastro@mgh.harvard.edu} (\href{https://orcid.org/0000-0001-7390-6354}{ORCID})

}
\keyword{internal}
