% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVrisk.R
\docType{package}
\name{CVrisk}
\alias{CVrisk}
\title{CVrisk: Compute Risk Scores for Cardiovascular Diseases}
\description{
This package calculate various cardiovascular disease risk scores from the
Framingham Heart Study, the American College of Cardiology, American Heart
Association and others.
}
\details{
Functions are provided to compute individual risk scores based on predictors
such as age, body mass index, HDL, total cholesterol and others.

A multiple scoring function also accepts a data frame and appends results
to score large datasets.
}
