% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frs_coef}
\alias{frs_coef}
\title{Model coefficients for ASCVD 10y FRS model}
\format{
A data frame with 2 obs. and 10 variables:
\describe{
  \item{gender}{Patient gender, either female or male}
  \item{ln_age}{Natural log of patient age}
  \item{ln_totchol}{Natural log of total cholesterol level}
  \item{ln_hdl}{Natural log of HDL level}
  \item{ln_untreated_sbp}{Natural log of untreated systolic blood pressure}
  \item{ln_treated_sbp}{Natural log of treated systolic blood pressure}
  \item{smoker}{Smoking status}
  \item{diabetes}{Diabetes status}
  \item{group_mean}{Grouped mean}
  \item{baseline_survival}{Baseline survival}
}
}
\usage{
frs_coef
}
\description{
A data set containing the Framingham risk score coefficients (full model
with lab features)
}
\references{
D’agostino, R.B., Vasan, R.S., Pencina, M.J., Wolf, P.A., Cobain, M.,
Massaro, J.M. and Kannel, W.B., 2008. General cardiovascular risk
profile for use in primary care. Circulation, 117(6), pp.743-753.
}
\keyword{datasets}
