% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample patient data}
\format{
A data frame with 3 obs. and 10 variables:
\describe{
  \item{age}{age in years}
  \item{gender}{Patient gender}
  \item{race}{race}
  \item{BMI}{Body mass index (kg/m2)}
  \item{sbp}{systolic blood pressure}
  \item{hdl}{HDL}
  \item{totchol}{Total cholesterol}
  \item{bp_med}{Patient is on blood pressure medication}
  \item{smoker}{Smoking status}
  \item{diabetes}{Diabetes status}
  \item{lipid_med}{Patient is on hyperlipidemic medication}
  \item{fh_heartattack}{Family history of heart attack}
  \item{cac}{Coronary artery calcification score}
}
}
\usage{
sample_data
}
\description{
A data set containing sample patient data
}
\keyword{datasets}
